Move Transmitter to Hono API #116
