+++ src/main/server/rpc/index.ts
@@ -4,6 +4,7 @@ import { serve, ServerType, type HttpBindings } from '@hono/node-server';
 import * as Server from './server/index.js';
 import * as Api from './api/index.js';
 import * as Admin from './admin/index.js';
+import * as Transmitter from './transmitter/index.js';
 
 const app = new Hono<{ Bindings: HttpBindings }>();
 let server: ServerType;
@@ -16,6 +17,9 @@ app.get('/', (c) => {
 app.route('/api', Api.get());
 app.route('/server', Server.get());
 app.route('/admin', Admin.get());
+if (alt.debug) {
+    app.route('/transmitter', Transmitter.get());
+}
 
 server = serve({ fetch: app.fetch, port: 8787 });
