+++ src/main/server/startup.ts
@@ -4,7 +4,6 @@ import '../translate/index.js';
 import { useTranslate } from '@Shared/translate.js';
 import { useConfig } from './config/index.js';
 import { useDatabase } from './database/index.js';
-import './systems/vscodeTransmitter.js';
 import './rpc/index.js';
 
 const config = useConfig();
